Fix: use qreal, not double in RouteSimulationPositionProviderPlugin
Summary: unbreaks build on platforms where qreal is float

Reviewers: rahn, nienhueser

Reviewed By: nienhueser

Differential Revision: https://phabricator.kde.org/D1431

+++ src/lib/marble/RouteSimulationPositionProviderPlugin.cpp
@@ -161,26 +161,26 @@ void RouteSimulationPositionProviderPlugin::update()
         if ( m_currentPosition.isValid() ) {
             //speed calculations
             //Max speed is set on points (m_lineStringInterpolated) based on formula. (max speed before points is calculated so the acceleration won't be exceeded)
-            const double acceleration = 1.5;
-            const double lookForwardDistance = 1000;
-            double checkedDistance = distanceSphere( m_currentPosition, m_lineStringInterpolated.at(m_currentIndex) )* m_marbleModel->planetRadius();
-            const double maxSpeed = 25;
-            const double minSpeed = 2;
-            double newSpeed = qMin((m_speed + acceleration*time), maxSpeed);
+            const qreal acceleration = 1.5;
+            const qreal lookForwardDistance = 1000;
+            qreal checkedDistance = distanceSphere( m_currentPosition, m_lineStringInterpolated.at(m_currentIndex) )* m_marbleModel->planetRadius();
+            const qreal maxSpeed = 25;
+            const qreal minSpeed = 2;
+            qreal newSpeed = qMin((m_speed + acceleration*time), maxSpeed);
             for (int i=qMax(1,m_currentIndex); i<m_lineStringInterpolated.size()-1 && checkedDistance<lookForwardDistance; ++i)
             {
-                double previousHeading = m_lineStringInterpolated.at( i-1 ).bearing( m_lineStringInterpolated.at( i ), GeoDataCoordinates::Degree, GeoDataCoordinates::FinalBearing );
-                double curveLength = 10;//we treat one point turn as a curve of length 10
-                double angleSum = 0;//sum of turn angles in a curve
+                qreal previousHeading = m_lineStringInterpolated.at( i-1 ).bearing( m_lineStringInterpolated.at( i ), GeoDataCoordinates::Degree, GeoDataCoordinates::FinalBearing );
+                qreal curveLength = 10;//we treat one point turn as a curve of length 10
+                qreal angleSum = 0;//sum of turn angles in a curve
                 for (int j=i+1; j<m_lineStringInterpolated.size() && curveLength<35; ++j)
                 {
-                    double newHeading = m_lineStringInterpolated.at( j-1 ).bearing( m_lineStringInterpolated.at( j ), GeoDataCoordinates::Degree, GeoDataCoordinates::FinalBearing );
-                    double differenceHeading = qAbs(previousHeading-newHeading);//angle of turn
+                    qreal newHeading = m_lineStringInterpolated.at( j-1 ).bearing( m_lineStringInterpolated.at( j ), GeoDataCoordinates::Degree, GeoDataCoordinates::FinalBearing );
+                    qreal differenceHeading = qAbs(previousHeading-newHeading);//angle of turn
                     if(differenceHeading>180) {
                         differenceHeading = 360 - differenceHeading;
                     }
                     angleSum +=differenceHeading;
-                    double maxSpeedAtTurn = qMax(((1 - (angleSum/60.0/curveLength*10.0))*maxSpeed), minSpeed);//speed limit at turn
+                    qreal maxSpeedAtTurn = qMax((1 - (static_cast<qreal>(angleSum/60.0/curveLength*10.0))*maxSpeed), minSpeed);//speed limit at turn
                     if( checkedDistance<25 && maxSpeedAtTurn<newSpeed )//if we are near turn don't accelerate, if we will have to slow down
                         newSpeed = qMin(newSpeed, qMax(m_speed,maxSpeedAtTurn));
                     // formulas:
@@ -190,9 +190,9 @@ void RouteSimulationPositionProviderPlugin::update()
                     // Vc = maxSpeedAtTurn
                     // s = checkedDistance
                     // a = acceleration
-                    double delta = maxSpeedAtTurn*maxSpeedAtTurn - 4.0*acceleration/2.0*(-checkedDistance);//delta = b*b-4*a*c
-                    double t = (-maxSpeedAtTurn+sqrt(delta))/(2.0*acceleration/2.0);//(-b+sqrt(delta))/(2*c)
-                    double maxCurrentSpeed = maxSpeedAtTurn + acceleration*t;
+                    qreal delta = maxSpeedAtTurn*maxSpeedAtTurn - 4.0*acceleration/2.0*(-checkedDistance);//delta = b*b-4*a*c
+                    qreal t = (-maxSpeedAtTurn+sqrt(delta))/(2.0*acceleration/2.0);//(-b+sqrt(delta))/(2*c)
+                    qreal maxCurrentSpeed = maxSpeedAtTurn + acceleration*t;
                     newSpeed = qMin(newSpeed, maxCurrentSpeed);
                     previousHeading = newHeading;
                     curveLength += distanceSphere( m_lineStringInterpolated.at( j-1 ), m_lineStringInterpolated.at( j ) )* m_marbleModel->planetRadius();
@@ -203,7 +203,7 @@ void RouteSimulationPositionProviderPlugin::update()
 
             //Assume the car's moving at m_speed m/s. The distance moved will be speed*time which is equal to the speed of the car if time is equal to one.
             //If the function isn't called once exactly after a second, multiplying by the time will compensate for the error and maintain the speed.
-            double fraction = m_speed*time/(distanceSphere( m_currentPosition, newPosition )* m_marbleModel->planetRadius());
+            qreal fraction = m_speed*time/(distanceSphere( m_currentPosition, newPosition )* m_marbleModel->planetRadius());
 
             //Interpolate and find the next point to move to if needed.
             if(fraction>0 && fraction <1){
